requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
groq==0.22.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.31
oauthlib==3.2.2
openai==1.72.0
orjson==3.10.16
packaging==24.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0
